+++ src/Components/PaymentMethod/PaymentMethodContainer.js
@@ -14,6 +14,7 @@ import {
   CREATE_PAYMENT,
   SET_FORMATTED_PRICE,
   SUBMIT_PAYMENT,
+  HANDLE_PAYPAL_SUBSCRIPTION,
   DISABLE_COUPON_BUTTON,
   APPLY_COUPON_CODE,
   SET_PERCENT_OFF,
@@ -32,6 +33,11 @@ import {
   showSuccess,
   hideError
 } from "../../utils/showing-error";
+import {
+  Subscription,
+  PaypalGateWay,
+  SUBSCRIPTION_TYPES
+} from "../../services/Subscription/Subscription.service";
 
 const initialState = {
   disableSubmit: false,
@@ -288,8 +294,65 @@ const PaymentMethodContainerWithoutStripe = ({
     return false;
   };
 
+  /**
+   * Handles subscriptions from PayPal gateway
+   * @param {string} paypalNonce
+   * @return {void}
+   */
+  const handlePaypalSubscription = (paypalNonce) => {
+    const subscription = new Subscription(new PaypalGateWay());
+
+    /**
+     * @TODO: Add flags for types instead of testing by properties
+     */
+
+    if (giftRecipient) {
+      return subscription.execute(
+        {
+          type: SUBSCRIPTION_TYPES.CREATE_GIFTED_SUBSCRIPTION,
+          token: paypalNonce,
+          plan,
+          couponCode,
+          product,
+          giftRecipient
+        },
+
+        (err, res) => {
+          dispatch({ type: DISABLE_SUBMIT, payload: false });
+
+          if (err) {
+            onFailure(err);
+            return displayError(getErrorMessages(err));
+          }
+
+          onSuccess(res);
+        }
+      );
+    }
+
+    return subscription.execute(
+      {
+        type: SUBSCRIPTION_TYPES.CREATE_SUBSCRIPTION,
+        token: paypalNonce,
+        plan,
+        couponCode,
+        product
+      },
+      (err, res) => {
+        dispatch({ type: DISABLE_SUBMIT, payload: false });
+
+        if (err) {
+          onFailure(err);
+          return displayError(getErrorMessages(err));
+        }
+
+        onSuccess(res);
+      }
+    );
+  };
+
   const purchase = (token, state, dispatch) => {
-    const order = state.order;
+    const { order } = state;
     order.email = window.Pelcro.user.read().email;
     const address = window.Pelcro.user.read().addresses
       ? window.Pelcro.user.read().addresses[
@@ -385,6 +448,11 @@ const PaymentMethodContainerWithoutStripe = ({
           return UpdateWithSideEffect(
             { ...state, disableSubmit: true },
             (state, dispatch) => submitPayment(state, dispatch)
+          );
+
+        case HANDLE_PAYPAL_SUBSCRIPTION:
+          return UpdateWithSideEffect(state, () =>
+            handlePaypalSubscription(action.payload)
           );
 
         case SET_FORMATTED_PRICE: